Move app/shared files to app/util
app/shared is not a concept for us
