+++ app/components/MessagePipeline/pauseFrameForPromise.test.ts
@@ -11,11 +11,11 @@
 //   found at http://www.apache.org/licenses/LICENSE-2.0
 //   You may not use this file except in compliance with the License.
 
-import delay from "@foxglove-studio/app/shared/delay";
-import signal from "@foxglove-studio/app/shared/signal";
+import delay from "@foxglove-studio/app/util/delay";
 import inAutomatedRunMode from "@foxglove-studio/app/util/inAutomatedRunMode";
 import { initializeLogEvent, resetLogEventForTests } from "@foxglove-studio/app/util/logEvent";
 import sendNotification from "@foxglove-studio/app/util/sendNotification";
+import signal from "@foxglove-studio/app/util/signal";
 
 import { pauseFrameForPromises, MAX_PROMISE_TIMEOUT_TIME_MS } from "./pauseFrameForPromise";
