+++ app/components/PanelToolbarInput.tsx
@@ -10,14 +10,10 @@
 //   This source code is licensed under the Apache License, Version 2.0,
 //   found at http://www.apache.org/licenses/LICENSE-2.0
 //   You may not use this file except in compliance with the License.
+
 import styled from "styled-components";
 
-export const PanelToolbarLabel = styled.label`
-  display: block;
-  padding-bottom: 5px;
-`;
-
-export const PanelToolbarInput = styled.input`
+export default styled.input`
   width: 100%;
   margin: 0;
 `;
