+++ app/dataProviders/CombinedDataProvider.test.ts
@@ -36,8 +36,8 @@ import RenameDataProvider from "@foxglove-studio/app/dataProviders/RenameDataPro
 import { mockExtensionPoint } from "@foxglove-studio/app/dataProviders/mockExtensionPoint";
 import { InitializationResult } from "@foxglove-studio/app/dataProviders/types";
 import { Bobject, BobjectMessage } from "@foxglove-studio/app/players/types";
-import delay from "@foxglove-studio/app/shared/delay";
 import { wrapJsObject } from "@foxglove-studio/app/util/binaryObjects";
+import delay from "@foxglove-studio/app/util/delay";
 import { SECOND_SOURCE_PREFIX } from "@foxglove-studio/app/util/globalConstants";
 import sendNotification from "@foxglove-studio/app/util/sendNotification";
 import { fromMillis } from "@foxglove-studio/app/util/time";
