+++ app/dataProviders/MemoryCacheDataProvider.test.ts
@@ -18,8 +18,8 @@ import MemoryDataProvider from "@foxglove-studio/app/dataProviders/MemoryDataPro
 import { CoreDataProviders } from "@foxglove-studio/app/dataProviders/constants";
 import { mockExtensionPoint } from "@foxglove-studio/app/dataProviders/mockExtensionPoint";
 import { Bobject, BobjectMessage, Message } from "@foxglove-studio/app/players/types";
-import delay from "@foxglove-studio/app/shared/delay";
 import { getObject } from "@foxglove-studio/app/util/binaryObjects";
+import delay from "@foxglove-studio/app/util/delay";
 import naturalSort from "@foxglove-studio/app/util/naturalSort";
 import sendNotification from "@foxglove-studio/app/util/sendNotification";
