+++ app/panels/ThreeDimensionalViz/CameraInfo/index.tsx
@@ -37,9 +37,9 @@ import {
   TargetPose,
 } from "@foxglove-studio/app/panels/ThreeDimensionalViz/threeDimensionalVizUtils";
 import { ThreeDimensionalVizConfig } from "@foxglove-studio/app/panels/ThreeDimensionalViz/types";
-import { point2DValidator, cameraStateValidator } from "@foxglove-studio/app/shared/validators";
 import colors from "@foxglove-studio/app/styles/colors.module.scss";
 import clipboard from "@foxglove-studio/app/util/clipboard";
+import { point2DValidator, cameraStateValidator } from "@foxglove-studio/app/util/validators";
 
 export const CAMERA_TAB_TYPE = "Camera";
