+++ app/panels/ThreeDimensionalViz/Interactions/Interaction.stories.tsx
@@ -23,12 +23,12 @@ import {
   POINT_CLOUD_WITH_ADDITIONAL_FIELDS,
 } from "@foxglove-studio/app/panels/ThreeDimensionalViz/commands/PointClouds/fixture/pointCloudData";
 import { MarkerStory } from "@foxglove-studio/app/panels/ThreeDimensionalViz/stories/MarkerStory";
-import delay from "@foxglove-studio/app/shared/delay";
-import tick from "@foxglove-studio/app/shared/tick";
 import PanelSetup, { triggerInputChange } from "@foxglove-studio/app/stories/PanelSetup";
 import { ScreenshotSizedContainer } from "@foxglove-studio/app/stories/storyHelpers";
 import colors from "@foxglove-studio/app/styles/colors.module.scss";
 import { simulateDragClick } from "@foxglove-studio/app/test/mouseEventsHelper";
+import delay from "@foxglove-studio/app/util/delay";
+import tick from "@foxglove-studio/app/util/tick";
 
 import Interactions, { OBJECT_TAB_TYPE, LINKED_VARIABLES_TAB_TYPE } from "./index";
 import useLinkedGlobalVariables from "./useLinkedGlobalVariables";
