+++ app/players/RandomAccessPlayer.test.ts
@@ -23,9 +23,9 @@ import {
   PlayerPresence,
   PlayerState,
 } from "@foxglove-studio/app/players/types";
-import delay from "@foxglove-studio/app/shared/delay";
-import signal from "@foxglove-studio/app/shared/signal";
+import delay from "@foxglove-studio/app/util/delay";
 import sendNotification from "@foxglove-studio/app/util/sendNotification";
+import signal from "@foxglove-studio/app/util/signal";
 import { fromNanoSec, getSeekToTime, SEEK_ON_START_NS } from "@foxglove-studio/app/util/time";
 
 import RandomAccessPlayer, {
