+++ app/players/automatedRun/AutomatedRunPlayer.test.ts
@@ -12,8 +12,8 @@
 //   You may not use this file except in compliance with the License.
 
 import { Progress } from "@foxglove-studio/app/players/types";
-import delay from "@foxglove-studio/app/shared/delay";
-import signal from "@foxglove-studio/app/shared/signal";
+import delay from "@foxglove-studio/app/util/delay";
+import signal from "@foxglove-studio/app/util/signal";
 
 import TestProvider from "../TestProvider";
 import AutomatedRunPlayer, {
