+++ app/util/CachedFilelike.test.ts
@@ -13,7 +13,7 @@
 
 import buffer from "buffer";
 
-import delay from "@foxglove-studio/app/shared/delay";
+import delay from "@foxglove-studio/app/util/delay";
 
 import CachedFilelike, { FileReader, FileStream } from "./CachedFilelike";
